Convert glRect to int16_t coordinates.

+++ include/graphics.h
@@ -98,7 +98,9 @@ void glHLine_(uint8_t x0, uint8_t x1, uint8_t y, uint8_t color);
 
 void glLine(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, uint8_t color);
 
-void glRect(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, uint8_t color);
+
+void glRect(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint8_t color);
+
 
 void glRectFill(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, uint8_t color);
